copilot issues
